# db: read path for values in value blocks