--- a/internal/base/lazy_value.go
+++ b/internal/base/lazy_value.go
@@ -163,7 +163,8 @@ type LazyFetcher struct {
 	// If the fetcher attempted to use buf *and* len(buf) was insufficient, it
 	// will allocate a new slice for the value. In either case it will set
 	// callerOwned to true.
-	ValueFetcher func(handle []byte, buf []byte) (val []byte, callerOwned bool, err error)
+	ValueFetcher func(
+		handle []byte, valLen int32, buf []byte) (val []byte, callerOwned bool, err error)
 	// Attribute includes the short attribute and value length.
 	Attribute   AttributeAndLen
 	fetched     bool
@@ -194,7 +195,8 @@ func (lv *LazyValue) fetchValue(buf []byte) (val []byte, callerOwned bool, err e
 	f := lv.Fetcher
 	if !f.fetched {
 		f.fetched = true
-		f.value, f.callerOwned, f.err = lv.Fetcher.ValueFetcher(lv.ValueOrHandle, buf)
+		f.value, f.callerOwned, f.err = lv.Fetcher.ValueFetcher(
+			lv.ValueOrHandle, lv.Fetcher.Attribute.ValueLen, buf)
 	}
 	return f.value, f.callerOwned, f.err
 }
@@ -226,17 +228,29 @@ func (lv *LazyValue) TryGetShortAttribute() (ShortAttribute, bool) {
 }
 
 // Clone creates a stable copy of the LazyValue, by appending bytes to buf.
+// The fetcher parameter must be non-nil and may be over-written and used
+// inside the returned LazyValue -- this is needed to avoid an allocation.
+// Most callers have at most K cloned LazyValues, where K is hard-coded, so
+// they can have a pool of exactly K LazyFetcher structs they can reuse in
+// these calls. The alternative of allocating LazyFetchers from a sync.Pool is
+// not viable since we have no code trigger for returning to the pool
+// (LazyValues are simply GC'd).
+//
 // REQUIRES: LazyValue.Value() has not been called.
 // Ensure that this is only called before LazyValue.Value is called. We don't
-// actually care if it was in-place but we don't want confusing things about whether
-// we need to clone the value inside the fetcher.
-func (lv *LazyValue) Clone(buf []byte) (LazyValue, []byte) {
+// actually care if it was in-place but we don't want to complicate behavior
+// regarding whether we need to clone the value inside the fetcher.
+func (lv *LazyValue) Clone(buf []byte, fetcher *LazyFetcher) (LazyValue, []byte) {
 	if invariants.Enabled && lv.Fetcher != nil && lv.Fetcher.fetched {
 		panic("value has already been fetched")
 	}
 	// INVARIANT: LazyFetcher.value is nil, so there is nothing to copy there.
 	vLen := len(lv.ValueOrHandle)
-	lvCopy := LazyValue{Fetcher: lv.Fetcher}
+	var lvCopy LazyValue
+	if lv.Fetcher != nil {
+		*fetcher = *lv.Fetcher
+		lvCopy.Fetcher = fetcher
+	}
 	if vLen == 0 {
 		return lvCopy, buf
 	}