--- a/sstable/testdata/prefixreader/iter
+++ b/sstable/testdata/prefixreader/iter
@@ -10,29 +10,29 @@ seek-prefix-ge a
 next
 next
 ----
-<a:1>
-<aa:2>
+<a:1>:A
+<aa:2>:AA
 .
 
 iter
 seek-prefix-ge aa
 next
 ----
-<aa:2>
+<aa:2>:AA
 .
 
 iter
 seek-prefix-ge aa
 prev
 ----
-<aa:2>
+<aa:2>:AA
 .
 
 iter
 seek-prefix-ge c
 prev
 ----
-<c:3>
+<c:3>:C
 .
 
 iter
@@ -44,14 +44,14 @@ iter
 seek-prefix-ge c
 next
 ----
-<c:3>
+<c:3>:C
 .
 
 iter
 seek-prefix-ge d
 next
 ----
-<d:4>
+<d:4>:D
 .
 
 iter
@@ -64,8 +64,8 @@ seek-prefix-ge c
 seek-prefix-ge d
 seek-prefix-ge e
 ----
-<c:3>
-<d:4>
+<c:3>:C
+<d:4>:D
 .
 
 iter
@@ -75,10 +75,10 @@ seek-prefix-ge a
 next
 next
 ----
-<c:3>
+<c:3>:C
 .
-<a:1>
-<aa:2>
+<a:1>:A
+<aa:2>:AA
 .
 
 iter
@@ -90,12 +90,12 @@ next
 seek-prefix-ge c
 next
 ----
-<aa:2>
+<aa:2>:AA
 .
-<a:1>
-<aa:2>
+<a:1>:A
+<aa:2>:AA
 .
-<c:3>
+<c:3>:C
 .
 
 iter
@@ -107,12 +107,12 @@ seek-prefix-ge a
 next
 next
 ----
-<c:3>
+<c:3>:C
 .
-<aa:2>
+<aa:2>:AA
 .
-<a:1>
-<aa:2>
+<a:1>:A
+<aa:2>:AA
 .
 
 iter
@@ -120,8 +120,8 @@ seek-prefix-ge a
 next
 next
 ----
-<a:1>
-<aa:2>
+<a:1>:A
+<aa:2>:AA
 .
 
 iter
@@ -130,16 +130,16 @@ next
 prev
 prev
 ----
-<a:1>
-<aa:2>
-<a:1>
+<a:1>:A
+<aa:2>:AA
+<a:1>:A
 .
 
 iter
 seek-prefix-ge a
 prev
 ----
-<a:1>
+<a:1>:A
 .
 
 iter
@@ -150,11 +150,11 @@ next
 next
 next
 ----
-<a:1>
-<a:1>
-<aa:2>
-<c:3>
-<d:4>
+<a:1>:A
+<a:1>:A
+<aa:2>:AA
+<c:3>:C
+<d:4>:D
 .
 
 iter
@@ -164,10 +164,10 @@ next
 next
 next
 ----
-<a:1>
-<aa:2>
-<c:3>
-<d:4>
+<a:1>:A
+<aa:2>:AA
+<c:3>:C
+<d:4>:D
 .
 
 iter
@@ -176,9 +176,9 @@ seek-ge c
 next
 next
 ----
-<aa:2>
-<c:3>
-<d:4>
+<aa:2>:AA
+<c:3>:C
+<d:4>:D
 .
 
 iter
@@ -188,10 +188,10 @@ next
 next
 next
 ----
-<aa:2>
-<aa:2>
-<c:3>
-<d:4>
+<aa:2>:AA
+<aa:2>:AA
+<c:3>:C
+<d:4>:D
 .
 
 iter
@@ -200,9 +200,9 @@ seek-lt c
 prev
 prev
 ----
-<aa:2>
-<aa:2>
-<a:1>
+<aa:2>:AA
+<aa:2>:AA
+<a:1>:A
 .
 
 iter
@@ -210,8 +210,8 @@ seek-lt c
 seek-prefix-ge aa
 prev
 ----
-<aa:2>
-<aa:2>
+<aa:2>:AA
+<aa:2>:AA
 .
 
 iter
@@ -221,9 +221,9 @@ next
 next
 
 ----
-<aa:2>
-<a:1>
-<aa:2>
+<aa:2>:AA
+<a:1>:A
+<aa:2>:AA
 .
 
 iter
@@ -233,9 +233,9 @@ next
 next
 
 ----
-<aa:2>
-<a:1>
-<aa:2>
+<aa:2>:AA
+<a:1>:A
+<aa:2>:AA
 .
 
 iter
@@ -263,8 +263,8 @@ seek-prefix-ge aa true
 seek-prefix-ge d true
 seek-prefix-ge c false
 ----
-<a:1>
-<a:1>
-<aa:2>
-<d:4>
-<c:3>
+<a:1>:A
+<a:1>:A
+<aa:2>:AA
+<d:4>:D
+<c:3>:C