There are a lot of trivial plumbing changes. The substantive changes are in sstable, specifically the new valueBlockReader, and how it interacts with the ReaderProvider in order to read values even when the sstable iterator may be closed. LazyValue.Clone now also takes a *LazyFetcher to avoid allocation.

The changes in blockIter are designed to minimize the performance loss for the v2 sstables, and for v3 sstables that don't have value blocks. Despite that, there is some performance loss in microbenchmarks (see details below). LevelIterSeqSeekPrefixGE/skip=4/use-next=true-16 shows a 10% slowdown from 185ns to 204ns on gceworker (on my macbook there is a smaller slowdown from 69ns to 72ns), for v2 on this PR, compared to master. There is a ~5% slowdown in BlockIterSeekGE/restart=16. The only new code involved here is two simple if-conditions: one for deciding we can construct an InPlaceValue and the second related to the first logic in SeekGE. Profiles confirm this and don't show enough in those places to explain the slow down -- it is possible that we are slightly worse with cache or register allocation behavior because of the larger structs, or maybe the if-blocks are causing hiccups in the instruction pipeline.

master vs PR (neither using v3):

```
name                                                                                        old time/op    new time/op    delta
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=false/with-tombstone=false-16     508ns ± 1%     509ns ± 0%     ~     (p=0.135 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=false/with-tombstone=true-16      421ns ± 0%     427ns ± 0%   +1.32%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=true/with-tombstone=false-16     1.11µs ± 1%    1.11µs ± 0%     ~     (p=0.968 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=true/with-tombstone=true-16       602ns ± 1%     603ns ± 1%     ~     (p=0.841 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=false/with-tombstone=false-16      594ns ± 0%     602ns ± 1%   +1.26%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=false/with-tombstone=true-16       444ns ± 0%     447ns ± 0%   +0.61%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=true/with-tombstone=false-16      1.14µs ± 0%    1.15µs ± 0%   +0.81%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=true/with-tombstone=true-16        609ns ± 2%     605ns ± 0%     ~     (p=0.690 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=false/with-tombstone=false-16     513ns ± 0%     512ns ± 1%     ~     (p=0.500 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=false/with-tombstone=true-16      428ns ± 0%     428ns ± 1%     ~     (p=0.810 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=true/with-tombstone=false-16     1.12µs ± 1%    1.11µs ± 0%   -0.61%  (p=0.048 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=true/with-tombstone=true-16       617ns ± 2%     611ns ± 1%     ~     (p=0.548 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=false/with-tombstone=false-16      599ns ± 0%     603ns ± 1%   +0.78%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=false/with-tombstone=true-16       450ns ± 0%     449ns ± 0%     ~     (p=0.151 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=true/with-tombstone=false-16      1.15µs ± 1%    1.15µs ± 1%     ~     (p=0.651 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=true/with-tombstone=true-16        621ns ± 1%     616ns ± 1%     ~     (p=0.310 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=false/with-tombstone=false-16     518ns ± 0%     517ns ± 1%     ~     (p=0.278 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=false/with-tombstone=true-16      434ns ± 1%     433ns ± 1%     ~     (p=0.548 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=true/with-tombstone=false-16     1.12µs ± 1%    1.13µs ± 1%     ~     (p=0.206 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=true/with-tombstone=true-16       621ns ± 1%     615ns ± 2%     ~     (p=0.310 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=false/with-tombstone=false-16      610ns ± 1%     612ns ± 1%     ~     (p=0.246 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=false/with-tombstone=true-16       459ns ± 1%     452ns ± 1%   -1.43%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=true/with-tombstone=false-16      1.15µs ± 1%    1.16µs ± 1%     ~     (p=0.095 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=true/with-tombstone=true-16        610ns ± 2%     615ns ± 1%     ~     (p=0.246 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=false/with-tombstone=false-16        635ns ± 0%     625ns ± 1%   -1.54%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=false/with-tombstone=true-16         523ns ± 0%     520ns ± 0%   -0.49%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=true/with-tombstone=false-16        1.36µs ± 0%    1.37µs ± 1%   +1.42%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=true/with-tombstone=true-16          855ns ± 0%     859ns ± 1%     ~     (p=0.151 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=false/with-tombstone=false-16         725ns ± 0%     717ns ± 1%   -1.02%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=false/with-tombstone=true-16          549ns ± 0%     544ns ± 0%   -0.85%  (p=0.016 n=5+4)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=true/with-tombstone=false-16         1.40µs ± 1%    1.39µs ± 1%     ~     (p=0.690 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=true/with-tombstone=true-16           877ns ± 1%     886ns ± 0%   +1.00%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=false/with-tombstone=false-16        672ns ± 1%     663ns ± 0%   -1.39%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=false/with-tombstone=true-16         562ns ± 0%     563ns ± 0%     ~     (p=0.810 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=true/with-tombstone=false-16        1.40µs ± 0%    1.42µs ± 0%   +1.03%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=true/with-tombstone=true-16          902ns ± 1%     910ns ± 1%   +0.85%  (p=0.032 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=false/with-tombstone=false-16         765ns ± 1%     761ns ± 1%     ~     (p=0.151 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=false/with-tombstone=true-16          587ns ± 0%     584ns ± 1%   -0.48%  (p=0.024 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=true/with-tombstone=false-16         1.44µs ± 0%    1.45µs ± 1%   +0.85%  (p=0.016 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=true/with-tombstone=true-16           922ns ± 1%     936ns ± 1%   +1.56%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=false/with-tombstone=false-16        746ns ± 1%     752ns ± 1%   +0.89%  (p=0.048 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=false/with-tombstone=true-16         632ns ± 1%     643ns ± 0%   +1.65%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=true/with-tombstone=false-16        1.49µs ± 1%    1.51µs ± 1%   +1.32%  (p=0.016 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=true/with-tombstone=true-16          983ns ± 0%     994ns ± 1%   +1.12%  (p=0.016 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=false/with-tombstone=false-16         840ns ± 0%     848ns ± 0%   +1.04%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=false/with-tombstone=true-16          658ns ± 0%     663ns ± 0%   +0.80%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=true/with-tombstone=false-16         1.54µs ± 0%    1.55µs ± 1%   +0.83%  (p=0.016 n=5+5)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=true/with-tombstone=true-16          1.01µs ± 0%    1.03µs ± 1%   +2.09%  (p=0.008 n=5+5)
IteratorSeqSeekGEWithBounds/two-level=false-16                                                 783ns ± 2%     776ns ± 0%     ~     (p=0.190 n=5+4)
IteratorSeqSeekGEWithBounds/two-level=true-16                                                  773ns ± 0%     776ns ± 1%     ~     (p=0.071 n=5+5)
IteratorSeekGENoop/withLimit=false-16                                                         58.2ns ± 2%    57.3ns ± 1%   -1.53%  (p=0.016 n=5+5)
IteratorSeekGENoop/withLimit=true-16                                                          98.6ns ± 0%    98.9ns ± 0%   +0.31%  (p=0.008 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/forward-16                                   5.35ms ± 1%    5.33ms ± 3%     ~     (p=0.548 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/backward-16                                  6.82ms ± 1%    6.91ms ± 4%     ~     (p=0.151 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/forward-16                                   4.85ms ± 4%    4.84ms ± 2%     ~     (p=0.548 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/backward-16                                  5.97ms ± 2%    6.02ms ± 2%     ~     (p=0.548 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/forward-16                                   2.66ms ± 0%    2.65ms ± 1%     ~     (p=0.190 n=4+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/backward-16                                  3.53ms ± 0%    3.50ms ± 2%     ~     (p=0.730 n=4+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/forward-16                                   361µs ± 1%     366µs ± 1%   +1.24%  (p=0.008 n=5+5)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/backward-16                                  704µs ± 1%     727µs ± 0%   +3.36%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=1,key-types=points-only-16                                        11.5µs ±15%    11.1µs ± 3%     ~     (p=1.000 n=5+5)
IteratorScan/keys=100,r-amp=1,key-types=points-and-ranges-16                                  11.7µs ± 5%    12.2µs ± 7%     ~     (p=0.421 n=5+5)
IteratorScan/keys=100,r-amp=3,key-types=points-only-16                                        19.4µs ± 5%    19.6µs ± 5%     ~     (p=0.421 n=5+5)
IteratorScan/keys=100,r-amp=3,key-types=points-and-ranges-16                                  21.0µs ± 5%    21.8µs ± 8%     ~     (p=0.310 n=5+5)
IteratorScan/keys=100,r-amp=7,key-types=points-only-16                                        29.7µs ± 5%    28.4µs ± 5%     ~     (p=0.095 n=5+5)
IteratorScan/keys=100,r-amp=7,key-types=points-and-ranges-16                                  30.6µs ±11%    31.3µs ±10%     ~     (p=0.548 n=5+5)
IteratorScan/keys=100,r-amp=10,key-types=points-only-16                                       35.8µs ± 8%    35.9µs ± 8%     ~     (p=0.690 n=5+5)
IteratorScan/keys=100,r-amp=10,key-types=points-and-ranges-16                                 34.5µs ± 5%    34.6µs ± 8%     ~     (p=1.000 n=5+5)
IteratorScan/keys=1000,r-amp=1,key-types=points-only-16                                       85.9µs ± 0%    87.2µs ± 1%   +1.50%  (p=0.029 n=4+4)
IteratorScan/keys=1000,r-amp=1,key-types=points-and-ranges-16                                  100µs ±18%      95µs ±13%     ~     (p=1.000 n=5+5)
IteratorScan/keys=1000,r-amp=3,key-types=points-only-16                                        151µs ± 1%     151µs ± 1%     ~     (p=0.556 n=4+5)
IteratorScan/keys=1000,r-amp=3,key-types=points-and-ranges-16                                  155µs ± 1%     155µs ± 1%     ~     (p=0.905 n=5+4)
IteratorScan/keys=1000,r-amp=7,key-types=points-only-16                                        197µs ± 3%     196µs ± 1%     ~     (p=0.548 n=5+5)
IteratorScan/keys=1000,r-amp=7,key-types=points-and-ranges-16                                  199µs ± 0%     201µs ± 1%   +0.87%  (p=0.032 n=5+4)
IteratorScan/keys=1000,r-amp=10,key-types=points-only-16                                       216µs ± 3%     217µs ± 3%     ~     (p=0.548 n=5+5)
IteratorScan/keys=1000,r-amp=10,key-types=points-and-ranges-16                                 223µs ± 3%     222µs ± 2%     ~     (p=0.905 n=5+4)
IteratorScan/keys=10000,r-amp=1,key-types=points-only-16                                       855µs ± 3%     869µs ± 2%     ~     (p=0.095 n=5+5)
IteratorScan/keys=10000,r-amp=1,key-types=points-and-ranges-16                                 881µs ± 0%     917µs ± 3%   +4.09%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=3,key-types=points-only-16                                      1.47ms ± 2%    1.46ms ± 2%     ~     (p=0.095 n=5+5)
IteratorScan/keys=10000,r-amp=3,key-types=points-and-ranges-16                                1.53ms ± 3%    1.50ms ± 1%   -2.32%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=7,key-types=points-only-16                                      1.88ms ± 1%    1.87ms ± 1%     ~     (p=0.421 n=5+5)
IteratorScan/keys=10000,r-amp=7,key-types=points-and-ranges-16                                1.90ms ± 0%    1.91ms ± 3%     ~     (p=0.413 n=4+5)
IteratorScan/keys=10000,r-amp=10,key-types=points-only-16                                     2.01ms ± 0%    2.03ms ± 1%   +0.91%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=10,key-types=points-and-ranges-16                               2.07ms ± 1%    2.07ms ± 1%     ~     (p=1.000 n=5+5)
CombinedIteratorSeek/range-key=false/batch=false-16                                           5.74µs ± 1%    5.74µs ± 1%     ~     (p=1.000 n=5+5)
CombinedIteratorSeek/range-key=false/batch=true-16                                            6.48µs ± 1%    6.56µs ± 1%   +1.28%  (p=0.008 n=5+5)
CombinedIteratorSeek/range-key=true/batch=false-16                                            13.7µs ± 1%    13.8µs ± 1%     ~     (p=0.151 n=5+5)
CombinedIteratorSeek/range-key=true/batch=true-16                                             14.5µs ± 0%    14.7µs ± 1%   +1.75%  (p=0.008 n=5+5)
CombinedIteratorSeek_Bounded-16                                                               87.6µs ± 1%    88.5µs ± 1%   +1.03%  (p=0.016 n=5+5)
CombinedIteratorSeekPrefix-16                                                                  179µs ± 0%     183µs ± 1%   +2.22%  (p=0.008 n=5+5)
LevelIterSeekGE/restart=16/count=5-16                                                         1.77µs ± 1%    1.89µs ± 5%   +6.87%  (p=0.008 n=5+5)
LevelIterSeqSeekGEWithBounds/restart=16/count=5-16                                             123ns ± 1%     123ns ± 0%     ~     (p=0.278 n=5+5)
LevelIterSeqSeekPrefixGE/skip=1/use-next=false-16                                              690ns ± 2%     691ns ± 0%     ~     (p=0.690 n=5+5)
LevelIterSeqSeekPrefixGE/skip=1/use-next=true-16                                              78.8ns ± 0%    82.4ns ± 0%   +4.59%  (p=0.008 n=5+5)
LevelIterSeqSeekPrefixGE/skip=2/use-next=false-16                                              700ns ± 1%     708ns ± 1%     ~     (p=0.175 n=5+5)
LevelIterSeqSeekPrefixGE/skip=2/use-next=true-16                                               113ns ± 1%     118ns ± 1%   +4.44%  (p=0.008 n=5+5)
LevelIterSeqSeekPrefixGE/skip=4/use-next=false-16                                              730ns ± 1%     730ns ± 1%     ~     (p=0.841 n=5+5)
LevelIterSeqSeekPrefixGE/skip=4/use-next=true-16                                               185ns ± 0%     204ns ± 1%  +10.13%  (p=0.008 n=5+5)
LevelIterSeqSeekPrefixGE/skip=8/use-next=false-16                                              856ns ± 1%     861ns ± 0%     ~     (p=0.421 n=5+5)
LevelIterSeqSeekPrefixGE/skip=8/use-next=true-16                                               954ns ± 1%     976ns ± 1%   +2.31%  (p=0.008 n=5+5)
LevelIterSeqSeekPrefixGE/skip=16/use-next=false-16                                             902ns ± 2%     900ns ± 4%     ~     (p=0.841 n=5+5)
LevelIterSeqSeekPrefixGE/skip=16/use-next=true-16                                              967ns ± 1%     971ns ± 3%     ~     (p=0.690 n=5+5)
LevelIterNext/restart=16/count=5-16                                                           24.4ns ± 1%    25.4ns ± 1%   +3.86%  (p=0.008 n=5+5)
LevelIterPrev/restart=16/count=5-16                                                           46.5ns ± 0%    46.7ns ± 0%     ~     (p=0.151 n=5+5)
MergingIterSeekGE/restart=16/count=1-16                                                       1.18µs ± 1%    1.20µs ± 2%     ~     (p=0.087 n=5+5)
MergingIterSeekGE/restart=16/count=2-16                                                       2.30µs ± 1%    2.34µs ± 1%   +1.75%  (p=0.016 n=5+5)
MergingIterSeekGE/restart=16/count=3-16                                                       3.53µs ± 2%    3.58µs ± 2%     ~     (p=0.151 n=5+5)
MergingIterSeekGE/restart=16/count=4-16                                                       4.81µs ± 1%    4.97µs ± 2%   +3.30%  (p=0.008 n=5+5)
MergingIterSeekGE/restart=16/count=5-16                                                       6.23µs ± 1%    6.37µs ± 2%   +2.18%  (p=0.032 n=5+5)
MergingIterNext/restart=16/count=1-16                                                         46.0ns ± 1%    47.6ns ± 1%   +3.40%  (p=0.008 n=5+5)
MergingIterNext/restart=16/count=2-16                                                         74.4ns ± 1%    77.2ns ± 3%   +3.71%  (p=0.008 n=5+5)
MergingIterNext/restart=16/count=3-16                                                         92.2ns ± 0%    95.3ns ± 0%   +3.38%  (p=0.008 n=5+5)
MergingIterNext/restart=16/count=4-16                                                          106ns ± 0%     109ns ± 0%   +2.82%  (p=0.008 n=5+5)
MergingIterNext/restart=16/count=5-16                                                          124ns ± 1%     129ns ± 2%   +3.36%  (p=0.008 n=5+5)
MergingIterPrev/restart=16/count=1-16                                                         66.8ns ± 1%    69.2ns ± 2%   +3.57%  (p=0.008 n=5+5)
MergingIterPrev/restart=16/count=2-16                                                         96.1ns ± 0%    99.3ns ± 1%   +3.31%  (p=0.008 n=5+5)
MergingIterPrev/restart=16/count=3-16                                                          115ns ± 0%     117ns ± 0%   +2.05%  (p=0.008 n=5+5)
MergingIterPrev/restart=16/count=4-16                                                          128ns ± 1%     129ns ± 1%   +1.13%  (p=0.048 n=5+5)
MergingIterPrev/restart=16/count=5-16                                                          144ns ± 1%     145ns ± 1%     ~     (p=0.087 n=5+5)
MergingIterSeqSeekGEWithBounds/levelCount=5-16                                                 613ns ± 1%     605ns ± 0%   -1.35%  (p=0.008 n=5+5)
MergingIterSeqSeekPrefixGE/skip=1/use-next=false-16                                           1.92µs ± 2%    1.93µs ± 0%     ~     (p=0.151 n=5+5)
MergingIterSeqSeekPrefixGE/skip=1/use-next=true-16                                             501ns ± 0%     499ns ± 1%     ~     (p=0.333 n=5+5)
MergingIterSeqSeekPrefixGE/skip=2/use-next=false-16                                           1.92µs ± 1%    1.95µs ± 0%   +1.64%  (p=0.008 n=5+5)
MergingIterSeqSeekPrefixGE/skip=2/use-next=true-16                                             536ns ± 1%     537ns ± 1%     ~     (p=0.548 n=5+5)
MergingIterSeqSeekPrefixGE/skip=4/use-next=false-16                                           1.97µs ± 0%    2.00µs ± 1%   +1.73%  (p=0.008 n=5+5)
…